Add detection for garbage in kind 10002
Some clients put wrong relays in kind 10002, or don't explain to users what to put in there. Probably because its still early and people don't know how to use outbox yet. So to detect useless kind 10002s we have some hardcoded relays, if these are found we can assume the kind 10002 is incorrectly configured and just ignore it these are relays that others are not supposed to read from, but announced in the kind 10002 as if it should